refactor(threads): useQueuedSend 二期拆 drain effects 与 fusion 子 hook
- 2287→1164 行：useQueueDrainEffects 916（9 个 drain effect+独占 state）+ useQueuedFusion 493（fuse 系 2 回调+4 effect）
- 参数对象每渲染新建禁缓存防 stale closure；deps 数组逐字；return 对象 10 键逐字不变
- 搬运块与 HEAD 逐字节一致（脚本断言）
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；useQueuedSend.test.tsx 77 passed ×2；gate 绿
（子代理 QueuedSendPhase2 执行，主代理复核验证）
