+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T16:56:46.571Z",
+  "generatedAt": "2026-08-30T16:58:02.035Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
